requirements.txt
discord.py
python-binance
openai
python-dotenv
pandas
numpy
google-genai
Pillow
